// src/FilterLists.Web/src/interfaces/List.ts
export interface List {
    id: number;
    chatUrl: string;
    description: string;
    descriptionSourceUrl: string;
    donateUrl: string;
    emailAddress: string;
    forumUrl: string;
    homeUrl: string;
    issuesUrl: string;
    languageIds: number[];
    licenseId: number;
    maintainerIds: number[];
    name: string;
    policyUrl: string;
    publishedDate: string;
    submissionUrl: string;
    syntaxId: number;
    tagIds: number[];
    viewUrl: string;
    viewUrlMirrors: string[];

    // auto-generated by useLists hook
    slug: string;
};